Rework ext.marshmallow.common.get_fields tests
