+++ tests/test_ext_marshmallow_common.py
@@ -1,12 +1,14 @@
 # -*- coding: utf-8 -*-
 import pytest
 
+from marshmallow import Schema, fields
+
 from apispec.ext.marshmallow.common import (
     make_schema_key,
     get_unique_schema_name,
     get_fields,
 )
-from .schemas import PetSchema, SampleSchema, ExcludeSchema, ExcludeSchema2
+from .schemas import PetSchema, SampleSchema
 
 
 class TestMakeSchemaKey:
@@ -49,19 +51,25 @@ class TestUniqueName:
         assert name_2 == "Pet2"
 
 
-class TestMetaExclude:
-    def test_meta_tuple_in_exclude(self):
-        keys = set(get_fields(ExcludeSchema).keys())
-        assert keys == {"field1", "field3", "field5"}
+class TestGetFields:
+    @pytest.mark.parametrize("exclude_type", (tuple, list))
+    @pytest.mark.parametrize("dump_only_type", (tuple, list))
+    def test_get_fields_meta_exclude_dump_only_as_list_and_tuple(
+        self, exclude_type, dump_only_type
+    ):
+        class ExcludeSchema(Schema):
+            field1 = fields.Int()
+            field2 = fields.Int()
+            field3 = fields.Int()
+            field4 = fields.Int()
+            field5 = fields.Int()
 
-    def test_meta_list_in_exclude(self):
-        keys = set(get_fields(ExcludeSchema2).keys())
-        assert keys == {"field1", "field2", "field4"}
+            class Meta:
+                ordered = True
+                exclude = exclude_type(("field1", "field2"))
+                dump_only = dump_only_type(("field3", "field4"))
 
-    def test_meta_tuple_in_dump_only(self):
-        keys = set(get_fields(ExcludeSchema, exclude_dump_only=True).keys())
-        assert keys == {"field1", "field5"}
-
-    def test_meta_list_in_dump_only(self):
-        keys = set(get_fields(ExcludeSchema2, exclude_dump_only=True).keys())
-        assert keys == {"field2", "field4"}
+        assert list(get_fields(ExcludeSchema).keys()) == ["field3", "field4", "field5"]
+        assert list(get_fields(ExcludeSchema, exclude_dump_only=True).keys()) == [
+            "field5"
+        ]
